Fixed bugs with reading articles
- EditedAt can be null
- StringTagConverter wasn't handling both formats of tags (list and string)
