--- src/DevTo.Api/Serialization/StringTagConverter.cs
@@ -1,28 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using Newtonsoft.Json;
-
-namespace DevTo.Api.Serialization
-{
-	public class StringTagConverter : JsonConverter
-	{
-		public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
-		{
-			throw new NotImplementedException("Unnecessary because this is read-only");
-		}
-
-		public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
-		{
-			return (reader.Value as string).Split(new[] { ',' }).Select(t => t.Trim());
-		}
-
-		public override bool CanWrite => false;
-
-		public override bool CanConvert(Type objectType)
-		{
-			return objectType == typeof(IEnumerable<string>);
-		}
-	}
-}
